+++ docs/changes.rst
@@ -10,6 +10,7 @@ v2.4.2 (UNRELEASED)
 
 **Crawlers**
 
+- New: ``bloomcounty2015``
 - Update: ``hjalmar`` to fix crash.